@@ -4,11 +4,12 @@ description: A sample command-line application.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 #  path: ^1.7.0
   path_to_regexp: ^0.4.0
+  flat: ^0.3.0
 
 dev_dependencies:
   pedantic: ^1.9.0
