@@ -1,17 +1,19 @@
 import 'dart:io';
+import 'package:flat/flat.dart';
 
 import 'package:drengr/router/router.dart';
 import 'package:drengr/container/container.dart';
 import 'package:drengr/config/config_reader.dart';
+
 
 class AppConfigurationException implements Exception {}
 
 class App {
 
   Router router;
-  Container container;
+  Container? container;
 
-  App({this.router, this.container});
+  App({required this.router, this.container});
 
   Future start() async {
     initializeContainer();
@@ -19,11 +21,7 @@ class App {
 
     router.container ??= container;
 
-    if (router != null) {
-      return await router.serveHTTP();
-    } else {
-      throw AppConfigurationException();
-    }
+    return await router.serveHTTP();
   }
 
   void initializeContainer () {
@@ -31,9 +29,12 @@ class App {
   }
 
   void loadConfigIntoContainer() {
-    var configFile = File('./config.yml');
-    var configReader = ConfigReader(file: configFile);
+    var configFile = File('config.yml');
+    var configReader = ConfigReader(file: configFile)..read();
     var config = configReader.parsed;
-    container?.bind('@config', (_) => config);
+    var flat = flatten(config);
+    flat.entries.forEach((element) {
+      container?.bind('@config.'+element.key, (_) => element.value);
+    });
   }
 }
