@@ -1,6 +1,3 @@
-import 'dart:io';
-
-import 'package:drengr/app/app.dart';
 import 'package:drengr/router/router.dart';
 import 'package:drengr/router/response.dart';
 import 'package:drengr/router/request.dart';
@@ -8,7 +5,7 @@ import 'package:drengr/router/request.dart';
 Future main() async {
   var router = Router();
   router.get('/hello', (Request request) {
-    return Response.Ok("Hello World!");
+    return Response.Ok('Hello World!');
   });
 
   router.get('/:name', (Request request) {
